+++ crates/cache/tests/engine_test.rs
@@ -35,9 +35,8 @@ mod create {
 
         assert!(dir.path().join(".moon/cache").exists());
         assert!(dir.path().join(".moon/cache/hashes").exists());
-        assert!(dir.path().join(".moon/cache/out").exists());
-        assert!(dir.path().join(".moon/cache/projects").exists());
-        assert!(dir.path().join(".moon/cache/tools").exists());
+        assert!(dir.path().join(".moon/cache/outputs").exists());
+        assert!(dir.path().join(".moon/cache/states").exists());
 
         dir.close().unwrap();
     }
@@ -56,7 +55,7 @@ mod delete_hash {
             .write_str("{}")
             .unwrap();
 
-        dir.child(".moon/cache/out/abc123.tar.gz")
+        dir.child(".moon/cache/outputs/abc123.tar.gz")
             .write_str("")
             .unwrap();
 
@@ -84,7 +83,7 @@ mod delete_hash {
             .write_str("{}")
             .unwrap();
 
-        dir.child(".moon/cache/out/abc123.tar.gz")
+        dir.child(".moon/cache/outputs/abc123.tar.gz")
             .write_str("")
             .unwrap();
 
@@ -114,7 +113,7 @@ mod delete_hash {
             .write_str("{}")
             .unwrap();
 
-        dir.child(".moon/cache/out/abc123.tar.gz")
+        dir.child(".moon/cache/outputs/abc123.tar.gz")
             .write_str("")
             .unwrap();
 
@@ -151,7 +150,7 @@ mod create_runfile {
         assert!(runfile.path.exists());
 
         assert_eq!(
-            fs::read_to_string(dir.path().join(".moon/cache/projects/123/runfile.json")).unwrap(),
+            fs::read_to_string(dir.path().join(".moon/cache/states/123/runfile.json")).unwrap(),
             "\"content\""
         );
 
@@ -180,7 +179,7 @@ mod cache_run_target_state {
     async fn loads_cache_if_it_exists() {
         let dir = assert_fs::TempDir::new().unwrap();
 
-        dir.child(".moon/cache/projects/foo/bar/lastRunState.json")
+        dir.child(".moon/cache/states/foo/bar/lastRun.json")
                 .write_str(r#"{"exitCode":123,"hash":"","lastRunTime":0,"stderr":"","stdout":"","target":"foo:bar"}"#)
                 .unwrap();
 
@@ -204,7 +203,7 @@ mod cache_run_target_state {
     async fn loads_cache_if_it_exists_and_cache_is_readonly() {
         let dir = assert_fs::TempDir::new().unwrap();
 
-        dir.child(".moon/cache/projects/foo/bar/lastRunState.json")
+        dir.child(".moon/cache/states/foo/bar/lastRun.json")
                 .write_str(r#"{"exitCode":123,"hash":"","lastRunTime":0,"stderr":"","stdout":"","target":"foo:bar"}"#)
                 .unwrap();
 
@@ -230,7 +229,7 @@ mod cache_run_target_state {
     async fn doesnt_load_if_it_exists_but_cache_is_off() {
         let dir = assert_fs::TempDir::new().unwrap();
 
-        dir.child(".moon/cache/projects/foo/bar/lastRunState.json")
+        dir.child(".moon/cache/states/foo/bar/lastRun.json")
                 .write_str(r#"{"exitCode":123,"hash":"","lastRunTime":0,"stderr":"","stdout":"","target":"foo:bar"}"#)
                 .unwrap();
 
@@ -327,8 +326,8 @@ mod cache_tool_state {
     async fn loads_cache_if_it_exists() {
         let dir = assert_fs::TempDir::new().unwrap();
 
-        dir.child(".moon/cache/tools/node-v1.2.3.json")
-            .write_str(r#"{"lastDepsInstallTime":123}"#)
+        dir.child(".moon/cache/states/toolNode-1.2.3.json")
+            .write_str(r#"{"lastVersionCheckTime":123}"#)
             .unwrap();
 
         let cache = CacheEngine::create(dir.path()).await.unwrap();
@@ -340,8 +339,7 @@ mod cache_tool_state {
         assert_eq!(
             item.item,
             ToolState {
-                last_deps_install_time: 123,
-                last_version_check_time: 0,
+                last_version_check_time: 123,
             }
         );
 
@@ -353,8 +351,8 @@ mod cache_tool_state {
     async fn loads_cache_if_it_exists_and_cache_is_readonly() {
         let dir = assert_fs::TempDir::new().unwrap();
 
-        dir.child(".moon/cache/tools/node-v4.5.6.json")
-            .write_str(r#"{"lastDepsInstallTime":123}"#)
+        dir.child(".moon/cache/states/toolNode-4.5.6.json")
+            .write_str(r#"{"lastVersionCheckTime":123}"#)
             .unwrap();
 
         let cache = CacheEngine::create(dir.path()).await.unwrap();
@@ -366,8 +364,7 @@ mod cache_tool_state {
         assert_eq!(
             item.item,
             ToolState {
-                last_deps_install_time: 123,
-                last_version_check_time: 0,
+                last_version_check_time: 123,
             }
         );
 
@@ -379,8 +376,8 @@ mod cache_tool_state {
     async fn doesnt_load_if_it_exists_but_cache_is_off() {
         let dir = assert_fs::TempDir::new().unwrap();
 
-        dir.child(".moon/cache/tools/system.json")
-            .write_str(r#"{"lastDepsInstallTime":123}"#)
+        dir.child(".moon/cache/states/tool-system.json")
+            .write_str(r#"{"lastVersionCheckTime":123}"#)
             .unwrap();
 
         let cache = CacheEngine::create(dir.path()).await.unwrap();
@@ -403,13 +400,13 @@ mod cache_tool_state {
             .await
             .unwrap();
 
-        item.item.last_deps_install_time = 123;
+        item.item.last_version_check_time = 123;
 
         run_with_env("", || item.save()).await.unwrap();
 
         assert_eq!(
             fs::read_to_string(item.path).unwrap(),
-            r#"{"lastDepsInstallTime":123,"lastVersionCheckTime":0}"#
+            r#"{"lastVersionCheckTime":123}"#
         );
 
         dir.close().unwrap();
@@ -441,7 +438,7 @@ mod cache_projects_state {
     async fn loads_cache_if_it_exists() {
         let dir = assert_fs::TempDir::new().unwrap();
 
-        dir.child(".moon/cache/projectsState.json")
+        dir.child(".moon/cache/states/projects.json")
             .write_str(r#"{"globs":["**/*"],"projects":{"foo":"bar"}}"#)
             .unwrap();
 
@@ -464,7 +461,7 @@ mod cache_projects_state {
     async fn loads_cache_if_it_exists_and_cache_is_readonly() {
         let dir = assert_fs::TempDir::new().unwrap();
 
-        dir.child(".moon/cache/projectsState.json")
+        dir.child(".moon/cache/states/projects.json")
             .write_str(r#"{"globs":["**/*"],"projects":{"foo":"bar"}}"#)
             .unwrap();
 
@@ -489,7 +486,7 @@ mod cache_projects_state {
     async fn doesnt_load_if_it_exists_but_cache_is_off() {
         let dir = assert_fs::TempDir::new().unwrap();
 
-        dir.child(".moon/cache/projectsState.json")
+        dir.child(".moon/cache/states/projects.json")
             .write_str(r#"{"globs":[],"projects":{"foo":"bar"}}"#)
             .unwrap();
 
@@ -508,14 +505,14 @@ mod cache_projects_state {
     async fn doesnt_load_if_it_exists_but_cache_is_stale() {
         let dir = assert_fs::TempDir::new().unwrap();
 
-        dir.child(".moon/cache/projectsState.json")
+        dir.child(".moon/cache/states/projects.json")
             .write_str(r#"{"globs":[],"projects":{"foo":"bar"}}"#)
             .unwrap();
 
         let now = to_millis(SystemTime::now()) - 100000;
 
         set_file_mtime(
-            dir.path().join(".moon/cache/projectsState.json"),
+            dir.path().join(".moon/cache/states/projects.json"),
             FileTime::from_unix_time((now / 1000) as i64, 0),
         )
         .unwrap();