+++ crates/lang-node/src/lib.rs
@@ -2,6 +2,7 @@ pub mod node;
 pub mod npm;
 pub mod package;
 pub mod pnpm;
+pub mod pnpm_workspace;
 pub mod tsconfig;
 pub mod yarn;
 pub mod yarn_classic;
@@ -28,7 +29,7 @@ pub const NPM: PackageManager = PackageManager {
 
 pub const PNPM: PackageManager = PackageManager {
     binary: "pnpm",
-    config_filenames: &[".npmrc", "pnpm-workspace.yaml", ".pnpmfile.cjs"],
+    config_filenames: &[".npmrc", ".pnpmfile.cjs", "pnpm-workspace.yaml"],
     default_version: "7.12.1",
     lock_filename: "pnpm-lock.yaml",
     manifest_filename: "package.json",