+++ crates/platform-node/src/actions/install_deps.rs
@@ -2,6 +2,7 @@ use moon_action::{Action, ActionContext, ActionStatus};
 use moon_config::NodePackageManager;
 use moon_contract::SupportedPlatform;
 use moon_error::map_io_to_fs_error;
+use moon_error::MoonError;
 use moon_lang::has_vendor_installed_dependencies;
 use moon_lang_node::{package::PackageJson, NODE, NPM};
 use moon_logger::{color, debug, warn};
@@ -69,16 +70,7 @@ fn add_engines_constraint(node: &NodeTool, package_json: &mut PackageJson) -> bo
     false
 }
 
-pub async fn install_deps(
-    _action: &mut Action,
-    context: &ActionContext,
-    workspace: Arc<RwLock<Workspace>>,
-    platform: &SupportedPlatform,
-) -> Result<ActionStatus, WorkspaceError> {
-    let workspace = workspace.read().await;
-    let node = workspace.toolchain.node.get()?;
-    let mut cache = workspace.cache.cache_tool_state(platform).await?;
-
+async fn sync_workspace(workspace: &Workspace, node: &NodeTool) -> Result<(), MoonError> {
     // Sync values to root `package.json`
     PackageJson::sync(&workspace.root, |package_json| {
         add_package_manager(node, package_json);
@@ -101,39 +93,76 @@ pub async fn install_deps(
         );
     }
 
-    // Get the last modified time of the root lockfile
-    let manager = node.get_package_manager();
-    let lockfile_name = manager.get_lock_filename();
-    let lockfile = workspace.root.join(&lockfile_name);
+    Ok(())
+}
+
+pub async fn install_deps(
+    _action: &mut Action,
+    context: &ActionContext,
+    workspace: Arc<RwLock<Workspace>>,
+    platform: &SupportedPlatform,
+    project_id: Option<&str>,
+) -> Result<ActionStatus, WorkspaceError> {
+    let workspace = workspace.read().await;
+    let node = workspace.toolchain.node.get_from_platform(platform)?;
+    let pm = node.get_package_manager();
+    let lock_filename = pm.get_lock_filename();
+    let manifest_filename = pm.get_manifest_filename();
+
+    // Determine the working directory and whether lockfiles and manifests have been modified
+    let working_dir;
+    let has_modified_files;
+
+    if let Some(project_id) = project_id {
+        let project = workspace.projects.load(project_id)?;
+
+        working_dir = project.root;
+        has_modified_files = context
+            .touched_files
+            .contains(&working_dir.join(&lock_filename))
+            || context
+                .touched_files
+                .contains(&working_dir.join(&manifest_filename));
+    } else {
+        working_dir = workspace.root.clone();
+        has_modified_files = context
+            .touched_files
+            .iter()
+            .any(|f| f.ends_with(&lock_filename) || f.ends_with(&manifest_filename));
+
+        // When installing deps in the workspace root, also sync applicable settings
+        sync_workspace(&workspace, node).await?;
+    }
+
+    // Install dependencies in the current project or workspace
+    let lock_filepath = working_dir.join(&lock_filename);
     let mut last_modified = 0;
+    let mut cache = workspace
+        .cache
+        .cache_deps_state(platform, project_id)
+        .await?;
 
-    if lockfile.exists() {
-        let lockfile_metadata = fs::metadata(&lockfile).await?;
-
+    if lock_filepath.exists() {
         last_modified = cache.to_millis(
-            lockfile_metadata
+            fs::metadata(&lock_filepath)
+                .await?
                 .modified()
-                .map_err(|e| map_io_to_fs_error(e, lockfile.clone()))?,
+                .map_err(|e| map_io_to_fs_error(e, lock_filepath.clone()))?,
         );
     }
 
-    // If a `package.json` has been modified manually, we should account for that
-    let has_modified_manifests = context
-        .touched_files
-        .iter()
-        .any(|f| f.ends_with(&NPM.manifest_filename) || f.ends_with(&lockfile_name));
-
-    // Install deps if the lockfile has been modified
-    // since the last time dependencies were installed!
-    if has_modified_manifests
-        || last_modified == 0
-        || last_modified > cache.item.last_deps_install_time
-    {
-        debug!(target: LOG_TARGET, "Installing Node.js dependencies");
+    // Install deps if the lockfile has been modified since the last time they were installed!
+    if has_modified_files || last_modified == 0 || last_modified > cache.item.last_install_time {
+        debug!(
+            target: LOG_TARGET,
+            "Installing {} dependencies in {}",
+            platform.label(),
+            color::path(&working_dir)
+        );
 
         // When in CI, we can avoid installing dependencies because
         // we can assume they've already been installed before moon runs!
-        if is_ci() && has_vendor_installed_dependencies(&workspace.root, &NODE) {
+        if is_ci() && has_vendor_installed_dependencies(&working_dir, &NODE) {
             warn!(
                 target: LOG_TARGET,
                 "In a CI environment and dependencies already exist, skipping install"
@@ -159,16 +188,16 @@ pub async fn install_deps(
 
         println!("{}", label_checkpoint(install_command, Checkpoint::Pass));
 
-        manager.install_dependencies(&workspace.toolchain).await?;
+        pm.install_dependencies(node, &working_dir).await?;
 
         if !is_ci() && node.config.dedupe_on_lockfile_change {
             debug!(target: LOG_TARGET, "Dedupeing dependencies");
 
-            manager.dedupe_dependencies(&workspace.toolchain).await?;
+            pm.dedupe_dependencies(node, &working_dir).await?;
         }
 
         // Update the cache with the timestamp
-        cache.item.last_deps_install_time = cache.now_millis();
+        cache.item.last_install_time = cache.now_millis();
         cache.save().await?;
 
         return Ok(ActionStatus::Passed);